Reorganized text handling
